removed Intro and Previously On
