+++ zkmarek/video/main.py
@@ -47,8 +47,6 @@ from zkmarek.video.slides.teaser_reference import TeaserReference
 from zkmarek.video.slides.wallet import WalletSlide
 from zkmarek.video.utils import get_slides_from_names, get_deck_name
 from zkmarek.video.slides.episode2.schnorr import SchnorrSlide
-from zkmarek.video.slides.episode2.PreviouslyOn import PreviouslyOn
-from zkmarek.video.slides.episode2.Intro import Intro_ep2
 
 config.width = 16
 config.height = 9
@@ -66,7 +64,7 @@ TEASER = [
 ]
 
 EPISODE1 = [
-    Intro(),
+    Intro(sound="data/sound/episode1/s1.wav"),
     TitleSlide(
         "Elliptic curves",
         subtitle="in Ethereum wallet",
@@ -110,8 +108,6 @@ EPISODE1 = [
 ]
 
 EPISODE2 = [
-    Intro_ep2(sound="data/sound/episode2/intro.mp3"),
-    PreviouslyOn(),
     TitleSlide("Digital Signature Algorithm", pre_wait_time=1.5),
     DigitalSignatureInterface(),
     Signature(),
