+++ zkmarek/video/slides/intro.py
@@ -1,5 +1,13 @@
-from manim import (LEFT, ORIGIN, RIGHT, Group, ImageMobject, MoveToTarget,
-                   Text, rate_functions)
+from manim import (
+    LEFT,
+    ORIGIN,
+    RIGHT,
+    Group,
+    ImageMobject,
+    MoveToTarget,
+    Text,
+    rate_functions,
+)
 
 from zkmarek.video.constant import PRIMARY_COLOR, PRIMARY_FONT
 from zkmarek.video.slides.common.slide_base import SlideBase
@@ -8,33 +16,32 @@ from zkmarek.video.slides.common.slide_base import SlideBase
 class Intro(SlideBase):
     marek: Text
     logo: ImageMobject
+    sound: str
 
-    def __init__(self):
+    def __init__(self, sound: str):
         super().__init__("Intro")
+        self.sound = sound
 
     def construct(self):
         self.logo = ImageMobject("data/brand/logo.png").scale(1.5)
-        self.marek = Text(
-            "Marek",
-            font_size=60,
-            font=PRIMARY_FONT,
-            color=PRIMARY_COLOR)
+        self.marek = Text("Marek", font_size=60, font=PRIMARY_FONT, color=PRIMARY_COLOR)
         self.marek.next_to(self.logo, RIGHT, buff=0.25)
         self.group = Group(self.logo, self.marek)
         self.group.move_to(ORIGIN)
 
     def animate_in(self, scene):
-        # self.play_music(scene, "data/sound/episode1/background.m4a")
-        self.new_subsection(scene, "Intro", sound="data/sound/episode1/s1.wav")
+        self.play_sound(scene, sound=self.sound)
         self.logo.generate_target()
         self.marek.generate_target()
         self.logo.shift(8 * LEFT)
         self.marek.shift(8 * RIGHT)
         scene.play(
-            MoveToTarget(self.marek, rate_func=rate_functions.ease_out_bounce,
-                run_time=2.25),
+            MoveToTarget(
+                self.marek, rate_func=rate_functions.ease_out_bounce, run_time=2.25
+            ),
             MoveToTarget(self.logo, rate_func=rate_functions.ease_out_bounce),
-                run_time=2.25)
+            run_time=2.25,
+        )
         scene.wait(4.5)
 
     def animate_out(self, scene):
@@ -43,4 +50,3 @@ class Intro(SlideBase):
         self.logo.generate_target()
         self.logo.target.shift(8 * LEFT)
         scene.play(MoveToTarget(self.marek), MoveToTarget(self.logo))
-
